requirements.txt
streamlit>=1.30.0
openai>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
rank_bm25>=0.2.2
langchain-text-splitters>=0.2.0
